# backend/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/projects/{id}:
    get:
      tags:
      - project-controller
      operationId: getById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Project"
    put:
      tags:
      - project-controller
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Project"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      tags:
      - project-controller
      operationId: delete_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /api/finance/{id}:
    put:
      tags:
      - finance-controller
      operationId: update_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceEntry"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FinanceEntry"
    delete:
      tags:
      - finance-controller
      operationId: delete_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /api/employees/{id}:
    get:
      tags:
      - employee-controller
      operationId: getEmployeeById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Employee"
    put:
      tags:
      - employee-controller
      operationId: updateEmployee
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Employee"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      tags:
      - employee-controller
      operationId: deleteEmployee
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/projects/{projectId}/allocations:
    get:
      tags:
      - project-allocation-controller
      operationId: list
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectAllocationDTO"
    post:
      tags:
      - project-allocation-controller
      operationId: add
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: stockId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: quantity
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectAllocationDTO"
...
  /api/finance:
    post:
      tags:
      - finance-controller
      operationId: create_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceEntry"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FinanceEntry"
...
  /api/employees:
    get:
      tags:
      - employee-controller
      operationId: getAllEmployees
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      tags:
      - employee-controller
      operationId: createEmployee
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Employee"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Employee"
...
  /api/finance/summary/{projectId}:
    get:
      tags:
      - finance-controller
      operationId: getSummary
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FinanceSummaryDTO"
  /api/finance/by-project/{projectId}:
    get:
      tags:
      - finance-controller
      operationId: getByProject
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinanceEntry"
...
  /api/projects/{projectId}/allocations/{allocationId}:
    delete:
      tags:
      - project-allocation-controller
      operationId: delete_7
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: allocationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
components:
  schemas:
    Client:
      type: object
      properties:
        clientId:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 100
          minLength: 0
        email:
          type: string
          maxLength: 100
          minLength: 0
        phone:
          type: string
          maxLength: 15
          minLength: 0
      required:
      - name
    Employee:
      type: object
      properties:
        employeeId:
          type: integer
          format: int64
        nic:
          type: string
          maxLength: 12
          minLength: 0
        hireDate:
          type: string
          format: date
        status:
          type: string
          maxLength: 20
          minLength: 0
        phone:
          type: string
          maxLength: 15
          minLength: 0
        role:
          type: string
          maxLength: 50
          minLength: 0
        name:
          type: string
          maxLength: 100
          minLength: 0
        address:
          type: string
          maxLength: 255
          minLength: 0
      required:
      - name
      - nic
    Project:
      type: object
      properties:
        projectId:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 150
          minLength: 0
        description:
          type: string
          maxLength: 1000
          minLength: 0
        location:
          type: string
          maxLength: 255
          minLength: 0
        budget:
          type: number
          minimum: 0.0
        status:
          type: string
          maxLength: 30
          minLength: 0
        startDate:
          type: string
          format: date
        plannedEndDate:
          type: string
          format: date
        createdDate:
          type: string
          format: date-time
        updatedDate:
          type: string
          format: date-time
        client:
          $ref: "#/components/schemas/Client"
      required:
      - client
      - name
...
    FinanceEntry:
      type: object
      properties:
        entryId:
          type: integer
          format: int64
        project:
          $ref: "#/components/schemas/Project"
        type:
          type: string
          maxLength: 30
          minLength: 0
        amount:
          type: number
          minimum: 0.0
        description:
          type: string
          maxLength: 500
          minLength: 0
        reference:
          type: string
          maxLength: 100
          minLength: 0
        entryDate:
          type: string
          format: date
      required:
      - amount
      - project
      - type
...
    ProjectAllocationDTO:
      type: object
      properties:
        allocationId:
          type: integer
          format: int64
        stockId:
          type: integer
          format: int64
        name:
          type: string
        resourceType:
          type: string
        unitOfMeasure:
          type: string
        quantity:
          type: integer
          format: int32
...
    FinanceSummaryDTO:
      type: object
      properties:
        projectId:
          type: integer
          format: int64
        budget:
          type: number
        totalCapital:
          type: number
        totalExpense:
          type: number
        balance:
          type: number